refactor: move API constants and drop status codes
